cambios de fechas a DateTimeLocal añadir cambios a tarifas nuevo curso...

+++ apps/actividadcargos/controller/form_1302.php
@@ -47,8 +47,8 @@ $Qpermiso = (integer) \filter_input(INPUT_POST,'permiso');
 
 $a_sel = (array)  \filter_input(INPUT_POST, 'sel', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 if (!empty($a_sel)) { //vengo de un checkbox
-	$Qid_item = strtok($a_sel[0],"#");
-	$eliminar =  strtok("#");
+	$Qid_item = (integer) strtok($a_sel[0],"#");
+	$eliminar = (integer) strtok("#");
 	// el scroll id es de la página anterior, hay que guardarlo allí
 	$oPosicion->addParametro('id_sel',$a_sel,1);
 	$scroll_id = (integer) \filter_input(INPUT_POST, 'scroll_id');
