+++ apps/actividadestudios/controller/form_1303.php
@@ -24,8 +24,8 @@ $Qid_asignatura = (integer) \filter_input(INPUT_POST, 'id_asignatura');
 
 $a_sel = (array)  \filter_input(INPUT_POST, 'sel', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 if (!empty($a_sel)) { //vengo de un checkbox
-	$Qid_activ = strtok($a_sel[0],"#");
-	$id_asignatura_real=strtok("#");
+    $Qid_activ = (integer) strtok($a_sel[0],"#");
+    $id_asignatura_real= (integer) strtok("#");
 	// el scroll id es de la página anterior, hay que guardarlo allí
 	$oPosicion->addParametro('id_sel',$a_sel,1);
 	$scroll_id = (integer) \filter_input(INPUT_POST, 'scroll_id');
