+++ apps/asistentes/controller/form_1301.php
@@ -46,8 +46,8 @@ $que_dl = (string)  \filter_input(INPUT_POST, 'que_dl');
 
 $a_sel = (array)  \filter_input(INPUT_POST, 'sel', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 if (!empty($a_sel)) { //vengo de un checkbox
-	$id_activ = strtok($a_sel[0],"#");
-	$id_asignatura=strtok("#");
+    $id_activ = (integer) strtok($a_sel[0],"#");
+    $id_asignatura= (integer) strtok("#");
 	// el scroll id es de la página anterior, hay que guardarlo allí
 	$oPosicion->addParametro('id_sel',$a_sel,1);
 	$scroll_id = (integer) \filter_input(INPUT_POST, 'scroll_id');
