+++ apps/devel/controller/modulos_ver.php
@@ -27,7 +27,7 @@ if ($Qmod == 'nuevo') {
 } else {
 	$a_sel = (array)  \filter_input(INPUT_POST, 'sel', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 	if (!empty($a_sel)) { //vengo de un checkbox
-		$Qid_mod = strtok($a_sel[0],"#");
+	    $Qid_mod = (integer) strtok($a_sel[0],"#");
 		// el scroll id es de la página anterior, hay que guardarlo allí
 		$oPosicion->addParametro('id_sel',$a_sel,1);
 		$scroll_id = (integer) \filter_input(INPUT_POST, 'scroll_id');
