+++ apps/usuarios/controller/role_grupmenu.php
@@ -17,7 +17,7 @@ $Qid_role = (string) \filter_input(INPUT_POST, 'id_role');
 
 $a_sel = (array)  \filter_input(INPUT_POST, 'sel', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 if (!empty($a_sel)) { //vengo de un checkbox
-	$Qid_role = strtok($a_sel[0],"#");
+    $Qid_role = (integer) strtok($a_sel[0],"#");
 	// el scroll id es de la página anterior, hay que guardarlo allí
 	$oPosicion->addParametro('id_sel',$a_sel,1);
 	$scroll_id = (integer) \filter_input(INPUT_POST, 'scroll_id');
