+++ apps/usuarios/controller/usuario_perm_menu.php
@@ -21,8 +21,8 @@ $Qquien = (string) \filter_input(INPUT_POST, 'quien');
 
 $a_sel = (array)  \filter_input(INPUT_POST, 'sel', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 if (!empty($a_sel)) { //vengo de un checkbox
-	$Qid_usuario = strtok($a_sel[0],"#");
-	$Qid_item=strtok("#");
+    $Qid_usuario = (integer) strtok($a_sel[0],"#");
+    $Qid_item= (integer) strtok("#");
 	// el scroll id es de la página anterior, hay que guardarlo allí
 	$oPosicion->addParametro('id_sel',$a_sel,1);
 	$scroll_id = (integer) \filter_input(INPUT_POST, 'scroll_id');
